Tests: Rename incorrect case `newBlockChainTests.md` file in repo
Commit 3d468a7 (`fixes path pointing to eth_tests`) renamed a test output
file subtly, changing just the capitalisation:

    - jsonTest(newFolder, "newBlockChainTests", testFixture, skipNewBCTests)
    + jsonTest(newFolder, "newBlockchainTests", testFixture, skipNewBCTests)

However, the file checked into the repo continued to be the old name, uppercase
`C`.  Of course, Mac and Windows have a sort of case insensitivity and Linux
does not.

This meant, on Linux, both files were created, test differences didn't make
themselves visible in `git diff`, and the repo would not get updates with
changed test output.  On Mac and Windows it worked.

Signed-off-by: Jamie Lokier <jamie@shareable.org>