Sync: Use stable nodekey/enode to accelerate sync testing
Make `run-nimbus-sync` look for and use `~/.nimbus/$TESTNET/nimbus/nodekey`
during Ethereum sync tests.  This is a private key which identifies the node.

If you have created that file, its contents should be a hex nodekey, same
format as Geth.  In fact you can use Geth to generate one.  If found,
`run-nimbus-sync` will use it as the nodekey, instead of Nimbus's default,
which is a random nodekey each time it is run.

Using the same nodekey for each run allows us to add the corresponding
`enode:...` URL (public key) as a trusted peer to the dedicated Geth instances,
using Geth's `admin.addTrustedPeer`.

This ensures Geth will almost always accept our connections, which is very
helpful for sync testing, instead of waiting a long time for a good peer.

Indeed, without this we might never get a willing good peer, due to reputation
effects while working on new sync methods.

Signed-off-by: Jamie Lokier <jamie@shareable.org>