Config: Add `--newsync` option and use it
This option enables new blockchain sync and real-time consensus algorithms that
will eventually replace the old, very limited sync.

New sync is work in progress.  It's included as an option rather than a code
branch, because it's more useful for testing this way, and must not conflict
anyway.  It's off by default.  Eventually this will become enabled by default
and the option will be removed.

Signed-off-by: Jamie Lokier <jamie@shareable.org>